+++ agent/impl/src/main/java/com/walmartlabs/concord/agent/ExecutionException.java
@@ -6,15 +6,7 @@ public class ExecutionException extends Exception {
         super(message);
     }
 
-    public ExecutionException(String message, Object[]... args) {
-        super(String.format(message, (Object) args));
-    }
-
     public ExecutionException(String message, Throwable cause) {
         super(message, cause);
-    }
-
-    public ExecutionException(String message, Throwable cause, Object[]... args) {
-        super(String.format(message, (Object) args), cause);
     }
 }