+++ agent/impl/src/main/java/com/walmartlabs/concord/agent/ExecutionManager.java
@@ -1,6 +1,5 @@
 package com.walmartlabs.concord.agent;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
 import com.walmartlabs.concord.agent.api.JobStatus;
@@ -17,9 +16,10 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.*;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Future;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.zip.ZipInputStream;
 import java.util.zip.ZipOutputStream;
@@ -30,48 +30,35 @@ public class ExecutionManager {
 
     private static final Logger log = LoggerFactory.getLogger(ExecutionManager.class);
     private static final long JOB_ENTRY_TTL = 8 * 60 * 60 * 1000; // 8 hours
-    private static final Collection<String> DEFAULT_JVM_ARGS = Arrays.asList(
-            "-Xmx128m",
-            "-Djavax.el.varArgs=true",
-            "-Djava.security.egd=file:/dev/./urandom",
-            "-Djava.net.preferIPv4Stack=true");
 
     private final Map<JobType, JobExecutor> jobExecutors;
-    private final ExecutorService executor;
     private final LogManager logManager;
     private final Configuration cfg;
-    private final DependencyManager dependencyManager;
 
-    private final Map<String, Future<?>> executions = new HashMap<>();
     private final Cache<String, JobStatus> statuses = CacheBuilder.newBuilder()
             .expireAfterAccess(JOB_ENTRY_TTL, TimeUnit.MILLISECONDS)
             .build();
-    private final Cache<String, Path> attachments = CacheBuilder.newBuilder()
+
+    private final Cache<String, JobInstance> instances = CacheBuilder.newBuilder()
             .expireAfterAccess(JOB_ENTRY_TTL, TimeUnit.MILLISECONDS)
             .build();
 
     private final Object mutex = new Object();
 
     @Inject
-    public ExecutionManager(@Named("executionPool") ExecutorService executor,
-                            JarJobExecutor jarJobExecutor,
+    public ExecutionManager(JarJobExecutor jarJobExecutor,
                             RunnerJobExecutor runnerJobExecutor,
                             LogManager logManager,
-                            Configuration cfg, DependencyManager dependencyManager) {
-
-        this.executor = executor;
+                            Configuration cfg) {
 
         this.logManager = logManager;
         this.cfg = cfg;
-        this.dependencyManager = dependencyManager;
         this.jobExecutors = new HashMap<>();
 
         jobExecutors.put(JobType.JAR, jarJobExecutor);
         jobExecutors.put(JobType.RUNNER, runnerJobExecutor);
-        jobExecutors.put(JobType.JUNIT_GROOVY, new JunitGroovyJobExecutor());
     }
 
-    @SuppressWarnings("unchecked")
     public void start(String instanceId, JobType type, String entryPoint, InputStream payload) throws ExecutionException {
         JobExecutor exec = jobExecutors.get(type);
         if (exec == null) {
@@ -81,79 +68,70 @@ public class ExecutionManager {
         // remove the previous log file
         // e.g. left after the execution was suspended
         logManager.delete(instanceId);
-
-        // create a log file synchronously
         logManager.touch(instanceId);
 
         Path tmpDir = extract(instanceId, payload);
-
-        Map<String, Object> agentParams = getAgentParameters(tmpDir);
-        Collection<String> jvmArgs = (Collection<String>) agentParams.getOrDefault(Constants.Agent.JVM_ARGS_KEY, DEFAULT_JVM_ARGS);
-        log.info("start ['{}'] -> JVM args: {}", instanceId, jvmArgs);
-
-        collectDependencies(instanceId, tmpDir);
 
         synchronized (mutex) {
             statuses.put(instanceId, JobStatus.RUNNING);
         }
 
-        Future<?> f = executor.submit(() -> {
-            try {
-                exec.exec(instanceId, tmpDir, entryPoint, jvmArgs);
-                synchronized (mutex) {
-                    statuses.put(instanceId, JobStatus.COMPLETED);
-                }
-            } catch (Exception e) {
-                log.error("start ['{}', {}, '{}'] -> failed", instanceId, type, entryPoint, e);
-                synchronized (mutex) {
-                    JobStatus s = statuses.getIfPresent(instanceId);
-                    if (s != JobStatus.CANCELLED) {
-                        statuses.put(instanceId, JobStatus.FAILED);
-                    }
-                }
-            } finally {
-                synchronized (ExecutionManager.this) {
-                    executions.remove(instanceId);
-                }
-            }
-
-            synchronized (mutex) {
-                Path attachmentsDir = tmpDir.resolve(Constants.Files.JOB_ATTACHMENTS_DIR_NAME);
-                if (Files.exists(attachmentsDir)) {
-                    attachments.put(instanceId, attachmentsDir);
-                }
-            }
-        });
+        JobInstance i;
+        try {
+            i = exec.start(instanceId, tmpDir, entryPoint);
+        } catch (Exception e) {
+            log.error("start ['{}', {}, '{}'] -> failed", instanceId, type, entryPoint, e);
+            handleError(instanceId, e);
+            throw e;
+        }
 
         synchronized (mutex) {
-            executions.put(instanceId, f);
+            instances.put(instanceId, i);
+        }
+
+        i.future().thenRun(() -> {
+            synchronized (mutex) {
+                statuses.put(instanceId, JobStatus.COMPLETED);
+            }
+        }).exceptionally(e -> {
+            handleError(instanceId, e);
+            return null;
+        });
+    }
+
+    private void handleError(String instanceId, Throwable e) {
+        synchronized (mutex) {
+            JobStatus s = statuses.getIfPresent(instanceId);
+            if (s != JobStatus.CANCELLED) {
+                statuses.put(instanceId, JobStatus.FAILED);
+            }
         }
     }
 
-    private void collectDependencies(String instanceId, Path tmpDir) throws ExecutionException {
-        Collection<String> deps = getDependencies(tmpDir);
-        if (deps != null && !deps.isEmpty()) {
-            logManager.log(instanceId, "Collecting dependencies...");
+    public int countRunning() {
+        int i = 0;
+        synchronized (mutex) {
+            for (Map.Entry<String, JobStatus> e : statuses.asMap().entrySet()) {
+                if (e.getValue() == JobStatus.RUNNING) {
+                    i++;
+                }
+            }
         }
-        try {
-            dependencyManager.collectDependencies(deps, tmpDir.resolve(Constants.Files.LIBRARIES_DIR_NAME));
-        } catch (IOException e) {
-            throw new ExecutionException("Error while collecting dependencies", e);
-        }
+        return i;
     }
 
     public void cancel() {
         Collection<String> ids;
         synchronized (mutex) {
-            ids = new ArrayList<>(executions.keySet());
+            ids = new ArrayList<>(instances.asMap().keySet());
         }
 
         for (String id : ids) {
-            cancel(id, false);
+            cancel(id);
         }
     }
 
-    public void cancel(String id, boolean waitToFinish) {
+    public void cancel(String id) {
         synchronized (mutex) {
             JobStatus s = statuses.getIfPresent(id);
             if (s != null && s == JobStatus.RUNNING) {
@@ -161,21 +139,12 @@ public class ExecutionManager {
             }
         }
 
-        Future<?> f = executions.get(id);
-        if (f == null) {
+        JobInstance i = instances.getIfPresent(id);
+        if (i == null) {
             return;
         }
-        f.cancel(true);
 
-        if (waitToFinish) {
-            while (!f.isDone()) {
-                try {
-                    Thread.sleep(1000);
-                } catch (InterruptedException e) {
-                    Thread.currentThread().interrupt();
-                }
-            }
-        }
+        i.kill();
     }
 
     public JobStatus getStatus(String id) {
@@ -191,18 +160,8 @@ public class ExecutionManager {
         return s;
     }
 
-    public int jobCount() {
-        synchronized (mutex) {
-            return executions.size();
-        }
-    }
-
     public Path zipAttachments(String id) throws IOException {
-        Path src;
-        synchronized (mutex) {
-            src = attachments.getIfPresent(id);
-        }
-
+        Path src = getAttachmentDir(id);
         if (src == null || !Files.exists(src)) {
             return null;
         }
@@ -216,16 +175,22 @@ public class ExecutionManager {
     }
 
     public void removeAttachments(String id) throws IOException {
-        Path p;
-        synchronized (mutex) {
-            p = attachments.getIfPresent(id);
-        }
-
+        Path p = getAttachmentDir(id);
         if (p == null || !Files.exists(p)) {
             return;
         }
 
         IOUtils.deleteRecursively(p);
+    }
+
+    private Path getAttachmentDir(String instanceId) {
+        synchronized (mutex) {
+            JobInstance i = instances.getIfPresent(instanceId);
+            if (i == null) {
+                return null;
+            }
+            return i.getWorkDir().resolve(Constants.Files.JOB_ATTACHMENTS_DIR_NAME);
+        }
     }
 
     private Path extract(String id, InputStream in) throws ExecutionException {
@@ -240,39 +205,6 @@ public class ExecutionManager {
             }
         } catch (IOException e) {
             throw new ExecutionException("Error while unpacking a payload", e);
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    private static Map<String, Object> getAgentParameters(Path payload) throws ExecutionException {
-        Path p = payload.resolve(Constants.Agent.AGENT_PARAMS_FILE_NAME);
-        if (!Files.exists(p)) {
-            return Collections.emptyMap();
-        }
-
-        try (InputStream in = Files.newInputStream(p)) {
-            ObjectMapper om = new ObjectMapper();
-            return om.readValue(in, Map.class);
-        } catch (IOException e) {
-            throw new ExecutionException("Error while reading an agent parameters file", e);
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    private static Collection<String> getDependencies(Path payload) throws ExecutionException {
-        Path p = payload.resolve(Constants.Files.REQUEST_DATA_FILE_NAME);
-        if (!Files.exists(p)) {
-            return Collections.emptyList();
-        }
-
-        try (InputStream in = Files.newInputStream(p)) {
-            ObjectMapper om = new ObjectMapper();
-            Map<String, Object> m = om.readValue(in, Map.class);
-
-            Collection<String> deps = (Collection<String>) m.get(Constants.Request.DEPENDENCIES_KEY);
-            return deps != null ? deps : Collections.emptyList();
-        } catch (IOException e) {
-            throw new ExecutionException("Error while reading a list of dependencies", e);
         }
     }
 }