+++ agent/impl/src/main/java/com/walmartlabs/concord/agent/JobExecutor.java
@@ -1,9 +1,8 @@
 package com.walmartlabs.concord.agent;
 
 import java.nio.file.Path;
-import java.util.Collection;
 
 public interface JobExecutor {
 
-    void exec(String id, Path workDir, String entryPoint, Collection<String> jvmArgs) throws ExecutionException;
+    JobInstance start(String instanceId, Path workDir, String entryPoint) throws ExecutionException;
 }