+++ agent/impl/src/main/java/com/walmartlabs/concord/agent/JobExecutorException.java
@@ -1,0 +1,12 @@
+package com.walmartlabs.concord.agent;
+
+public class JobExecutorException extends RuntimeException {
+
+    public JobExecutorException(String message) {
+        super(message);
+    }
+
+    public JobExecutorException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}