+++ agent/impl/src/main/java/com/walmartlabs/concord/agent/JobInstance.java
@@ -1,0 +1,13 @@
+package com.walmartlabs.concord.agent;
+
+import java.nio.file.Path;
+import java.util.concurrent.CompletableFuture;
+
+public interface JobInstance {
+
+    Path getWorkDir();
+
+    void kill();
+
+    CompletableFuture<?> future();
+}