+++ agent/impl/src/main/java/com/walmartlabs/concord/agent/RunnerJobExecutor.java
@@ -1,27 +1,28 @@
 package com.walmartlabs.concord.agent;
 
+import com.walmartlabs.concord.project.Constants;
+
 import javax.inject.Inject;
 import javax.inject.Named;
 import java.nio.file.Path;
-import java.util.Collection;
-import java.util.Collections;
+import java.util.concurrent.ExecutorService;
 
 @Named
-public class RunnerJobExecutor implements JobExecutor {
-
-    private final Configuration cfg;
-    private final JarJobExecutor jarExecutor;
+public class RunnerJobExecutor extends JarJobExecutor {
 
     @Inject
-    public RunnerJobExecutor(Configuration cfg, JarJobExecutor jarExecutor) {
-        this.cfg = cfg;
-        this.jarExecutor = jarExecutor;
+    public RunnerJobExecutor(Configuration cfg, LogManager logManager, DependencyManager dependencyManager, ExecutorService executorService) {
+        super(cfg, logManager, dependencyManager, executorService);
     }
 
     @Override
-    public void exec(String id, Path workDir, String entryPoint, Collection<String> jvmArgs) throws ExecutionException {
-        String mainClass = "com.walmartlabs.concord.runner.Main";
-        Collection<String> runnerCp = Collections.singleton(cfg.getRunnerPath().normalize().toString());
-        jarExecutor.exec(id, workDir, mainClass, jvmArgs, runnerCp);
+    protected String getMainClass(Path workDir, String entryPoint) throws ExecutionException {
+        return "com.walmartlabs.concord.runner.Main";
+    }
+
+    @Override
+    protected String createClassPath(String entryPoint) {
+        String runnerPath = getCfg().getRunnerPath().normalize().toString();
+        return Constants.Files.LIBRARIES_DIR_NAME + "/*:" + runnerPath;
     }
 }